src/requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
pillow
pandas
scikit-learn
matplotlib
python-multipart
